Move livecheck_strategy back into subdirectory
