+++ cmd/livecheck.rb
@@ -2,7 +2,7 @@
 
 require "cli/parser"
 
-require_relative "../livecheck_strategy"
+require_relative "../livecheck/livecheck_strategy"
 require_relative "../livecheck/extend/formulary"
 
 module Homebrew
