+++ livecheck/livecheck_strategy.rb
