+++ livecheck/livecheck_strategy/apache.rb
