+++ livecheck/livecheck_strategy/git.rb
