+++ livecheck/livecheck_strategy/gnome.rb
