+++ livecheck/livecheck_strategy/gnu.rb
