+++ livecheck/livecheck_strategy/hackage.rb
