+++ livecheck/livecheck_strategy/launchpad.rb
